Delete midenset

+++ PsychologicalSupports/PsychologicalSupports/Dependencies/Interface/IRepository.cs
@@ -1,15 +1,14 @@
 ﻿using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace PsychologicalSupports.Models.Dependencies
 {
     public interface IRepository<T> where T : class
     {
         IEnumerable<T> List();
-        T Get(int? id);
+        T Get(long? id);
         void Create(T item);
         void Edit(T item);
-        void Delete(int id);
+        void Delete(long id);
     }
-    
+
 }
